churn-app/pyproject.toml
[project]
name = "churn-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "~=3.12.0"
dependencies = [
    "churn-lib",
    "fastapi>=0.115.0",
    "langchain[openai]>=1.2.0",
    "langchain-openai>=1.1.5",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "uvicorn[standard]>=0.32.0",
]

[tool.uv.sources]
churn-lib = { path = "../churn-lib", editable = true }
